+++ intents/connectors/_experimental/snips/export.py
@@ -6,7 +6,9 @@ from intents.model.intent import _IntentMetaclass
 from intents.model.entity import _EntityMetaclass
 from intents.language import intent_language, entity_language
 from intents.language import agent_supported_languages, LanguageCode
+from intents.service_connector import PatchedEntityMapping
 from intents.connectors._experimental.snips import SnipsConnector
+from intents.connectors._experimental.snips import entities as snips_entities
 from intents.connectors._experimental.snips import agent_format as af
 
 def render(connector: SnipsConnector) -> Dict[LanguageCode, dict]:
@@ -19,10 +21,11 @@ def render(connector: SnipsConnector) -> Dict[LanguageCode, dict]:
 
 def render_dataset(connector: SnipsConnector, lang: LanguageCode) -> af.Dataset:
     entities = {e.name: render_entity(connector, e, lang) for e in connector.agent_cls._entities_by_name.values()}
-    entities.update({
-        'snips/musicArtist': {},
-        'snips/number': {}
-    }) # TODO: expand/read from snips.entities module
+    # TODO: only render those that are actually used in Agent
+    all_patched_entities = [m.builtin_entity for m in connector.entity_mappings.values() if isinstance(m, PatchedEntityMapping)]
+    entities.update({e.name: render_entity(connector, e, lang) for e in all_patched_entities})
+    all_builtin_entities = [e.value for e in snips_entities.BuiltinEntityTypes]
+    entities.update({e: {} for e in all_builtin_entities})
     return af.Dataset(
         intents={i.name: render_intent(connector, i, lang) for i in connector.agent_cls.intents},
         entities=entities,
@@ -44,9 +47,11 @@ def render_intent(
                 utterance_data.append(af.DatasetIntentUtteranceTextSegment(chunk.text))
             elif isinstance(chunk, intent_language.EntityUtteranceChunk):
                 chunk: intent_language.EntityUtteranceChunk
+                mapping = connector.entity_mappings.get(chunk.entity_cls)
+                entity_name = mapping.service_name if mapping else chunk.entity_cls.name
                 utterance_data.append(af.DatasetIntentUtteranceEntitySegment(
                     text=chunk.parameter_value,
-                    entity=chunk.entity_cls.name,
+                    entity=entity_name,
                     slot_name=chunk.parameter_name
                 ))
             else: